+++ packages/workdo/Account/src/Models/CustomerPayment.php
@@ -38,6 +38,11 @@ class CustomerPayment extends Model
         return $this->belongsTo(User::class, 'customer_id');
     }
 
+    public function customerDetails(): BelongsTo
+    {
+        return $this->belongsTo(Customer::class, 'customer_id', 'user_id');
+    }
+
     public function bankAccount(): BelongsTo
     {
         return $this->belongsTo(BankAccount::class, 'bank_account_id');